+++ src/resource_browser/hints.rs
@@ -12,6 +12,7 @@ const EXPAND_DATASTORE: &str = "d datastore";
 const EXPAND_HOST: &str = "h host";
 const EXPAND_VM: &str = "v vm";
 const EXPAND_TASK: &str = "t task";
+const VM_ACTIONS: &str = "x actions";
 
 const CLUSTER_EXPAND_HINTS: &[&str] = &[
     EXPAND_NETWORK,
@@ -24,7 +25,7 @@ const HOST_EXPAND_HINTS: &[&str] = &[EXPAND_NETWORK, EXPAND_DATASTORE, EXPAND_VM
 const DATASTORE_EXPAND_HINTS: &[&str] = &[EXPAND_HOST, EXPAND_VM, EXPAND_TASK];
 const NETWORK_EXPAND_HINTS: &[&str] = &[EXPAND_HOST, EXPAND_VM, EXPAND_TASK];
 
-const VM_EXPAND_HINTS: &[&str] = &[EXPAND_TASK];
+const VM_EXPAND_HINTS: &[&str] = &[VM_ACTIONS, EXPAND_TASK];
 pub(crate) fn get_expand_hint(resource_type: ResourceType) -> &'static [&'static str] {
     match resource_type {
         ResourceType::Cluster => CLUSTER_EXPAND_HINTS,